WIP: Issue 1738 (#3132)
Fixes #1738
Fixes #2698
Fixes #2210

Changes made in this Pull Request:
 - `ts.dimensions` gives a **view** of the box info
 - `atomgroup.dimensions` gives a **copy** of the box info
 - no longer use `ts._unitcell` any more (use `ts.dimensions`)
 - `ts.dimensions` can be `None` for no box info, was previously often all zeros, but not consistently.  This means blindly calling e.g. `distance_array(..., box=ag.dimensions)` will just work
 - Removes reader specific `Timestep` objects, now everything works from `base.Timestep`
