+++ package/CHANGELOG
@@ -185,6 +185,8 @@ Enhancements
     to transfer data from HDF5 file into timestep. (PR #3293)
 
 Changes
+  * Timestep now stores information in 'C' memory layout instead of
+    the previous 'F' default (PR #1738)
   * `analysis.rdf.InterRDF` and `analysis.rdf.InterRDF_s` now store
     their result attributes using the `analysis.base.Results` class
     (Issue #3276, #3261)
