+++ package/MDAnalysis/coordinates/PDBQT.py
@@ -166,12 +166,13 @@ class PDBQTReader(base.SingleFrameReaderBase):
         self.ts = self._Timestep.from_coordinates(
             coords,
             **self._ts_kwargs)
-        self.ts._unitcell[:] = unitcell
+        self.ts.dimensions = unitcell
         self.ts.frame = 0  # 0-based frame number
         if self.convert_units:
             # in-place !
             self.convert_pos_from_native(self.ts._pos)
-            self.convert_pos_from_native(self.ts._unitcell[:3])
+            if self.ts.dimensions is not None:
+                self.convert_pos_from_native(self.ts.dimensions[:3])
 
     def Writer(self, filename, **kwargs):
         """Returns a permissive (simple) PDBQTWriter for *filename*.
