+++ package/MDAnalysis/coordinates/XDR.py
@@ -164,7 +164,8 @@ class XDRBaseReader(base.ReaderBase):
         self.ts.dt = dt
         self.ts.dimensions = triclinic_box(*frame.box)
         if self.convert_units:
-            self.convert_pos_from_native(self.ts.dimensions[:3])
+            if self.ts.dimensions is not None:
+                self.convert_pos_from_native(self.ts.dimensions[:3])
 
     @classmethod
     def parse_n_atoms(cls, filename, **kwargs):
