+++ package/MDAnalysis/coordinates/base.py
@@ -42,7 +42,6 @@ MDAnalysis.
    .. automethod:: __init__
    .. automethod:: from_coordinates
    .. automethod:: from_timestep
-   .. automethod:: _init_unitcell
    .. autoattribute:: n_atoms
    .. attribute::`frame`
 
@@ -237,8 +236,10 @@ class Timestep(object):
     .. versionchanged:: 2.0.0
        Timestep now can be (un)pickled. Weakref for Reader
        will be dropped.
+       Timestep now stores in to numpy array memory in 'C' order rather than
+       'F' (Fortran).
     """
-    order = 'F'
+    order = 'C'
 
     def __init__(self, n_atoms, **kwargs):
         """Create a Timestep, representing a frame of a trajectory
@@ -292,6 +293,7 @@ class Timestep(object):
         self._has_positions = False
         self._has_velocities = False
         self._has_forces = False
+        self._has_dimensions = False
 
         # These will allocate the arrays if the has flag
         # gets set to True
@@ -299,7 +301,7 @@ class Timestep(object):
         self.has_velocities = kwargs.get('velocities', False)
         self.has_forces = kwargs.get('forces', False)
 
-        self._unitcell = self._init_unitcell()
+        self._unitcell = np.zeros(6, dtype=np.float32)
 
         # set up aux namespace for adding auxiliary data
         self.aux = Namespace()
@@ -400,11 +402,6 @@ class Timestep(object):
     def __setstate__(self, state):
         self.__dict__.update(state)
 
-    def _init_unitcell(self):
-        """Create custom datastructure for :attr:`_unitcell`."""
-        # override for other Timesteps
-        return np.zeros((6), np.float32)
-
     def __eq__(self, other):
         """Compare with another Timestep
 
@@ -423,6 +420,15 @@ class Timestep(object):
             return False
         if self.has_positions:
             if not (self.positions == other.positions).all():
+                return False
+
+        if self.dimensions is None:
+            if other.dimensions is not None:
+                return False
+        else:
+            if other.dimensions is None:
+                return False
+            if not (self.dimensions == other.dimensions).all():
                 return False
 
         if not self.has_velocities == other.has_velocities:
@@ -567,7 +573,7 @@ class Timestep(object):
             velocities=vel,
             forces=force)
 
-        new_TS._unitcell = self._unitcell.copy()
+        new_TS.dimensions = self.dimensions
 
         new_TS.frame = self.frame
 
@@ -781,28 +787,30 @@ class Timestep(object):
 
     @property
     def dimensions(self):
-        """unitcell dimensions (*A*, *B*, *C*, *alpha*, *beta*, *gamma*)
+        """View of unitcell dimensions (*A*, *B*, *C*, *alpha*, *beta*, *gamma*)
 
         lengths *a*, *b*, *c* are in the MDAnalysis length unit (Å), and
         angles are in degrees.
-
-        Setting dimensions will populate the underlying native format
-        description of box dimensions
         """
-        # The actual Timestep._unitcell depends on the underlying
-        # trajectory format. It can be e.g. six floats representing
-        # the box edges and angles or the 6 unique components of the
-        # box matrix or the full box matrix.
-        return self._unitcell
+        if (self._unitcell[:3] == 0).all():
+            return None
+        else:
+            return self._unitcell
 
     @dimensions.setter
     def dimensions(self, box):
-        self._unitcell[:] = box
+        if box is None:
+            self._unitcell[:] = 0
+        else:
+            self._unitcell[:] = box
 
     @property
     def volume(self):
         """volume of the unitcell"""
-        return core.box_volume(self.dimensions)
+        if self.dimensions is None:
+            return 0
+        else:
+            return core.box_volume(self.dimensions)
 
     @property
     def triclinic_dimensions(self):
@@ -840,7 +848,10 @@ class Timestep(object):
 
         .. versionadded:: 0.11.0
         """
-        return core.triclinic_vectors(self.dimensions)
+        if self.dimensions is None:
+            return None
+        else:
+            return core.triclinic_vectors(self.dimensions)
 
     @triclinic_dimensions.setter
     def triclinic_dimensions(self, new):
@@ -848,7 +859,10 @@ class Timestep(object):
 
         .. versionadded:: 0.11.0
         """
-        self.dimensions = core.triclinic_box(*new)
+        if new is None:
+            self.dimensions = None
+        else:
+            self.dimensions = core.triclinic_box(*new)
 
     @property
     def dt(self):
@@ -1615,7 +1629,7 @@ class ProtoReader(IOBase, metaclass=_Readermeta):
         # self._jump_to_frame(frame)
         # ts = self.ts
         # ts.frame = self._read_next_frame(ts._x, ts._y, ts._z,
-        #                                  ts._unitcell, 1)
+        #                                  ts.dimensions, 1)
         # return ts
 
     def _read_frame_with_aux(self, frame):
@@ -2207,7 +2221,10 @@ class WriterBase(IOBase, metaclass=_Writermeta):
         """
         # override if the native trajectory format does NOT use
         # [A,B,C,alpha,beta,gamma]
-        lengths, angles = ts.dimensions[:3], ts.dimensions[3:]
+        if ts.dimensions is None:
+            lengths, angles = np.zeros(3), np.zeros(3)
+        else:
+            lengths, angles = ts.dimensions[:3], ts.dimensions[3:]
         if not inplace:
             lengths = lengths.copy()
         lengths = self.convert_pos_to_native(lengths)
