+++ testsuite/MDAnalysisTests/coordinates/test_gro.py
@@ -26,7 +26,7 @@ from MDAnalysis.coordinates.GRO import GROReader, GROWriter
 from MDAnalysis.transformations import translate
 from MDAnalysisTests import make_Universe
 from MDAnalysisTests.coordinates.base import (
-    BaseReference, BaseReaderTest, BaseWriterTest, BaseTimestepTest,
+    BaseReference, BaseReaderTest, BaseWriterTest,
 )
 from MDAnalysisTests.coordinates.reference import RefAdK
 from MDAnalysisTests.datafiles import (
@@ -477,26 +477,6 @@ class TestGrowriterReindex(object):
                 line = grofile.readline()
             assert line.startswith('    2CL      CL    1')
 
-class TestGROTimestep(BaseTimestepTest):
-    Timestep = mda.coordinates.GRO.Timestep
-    name = "GRO"
-    has_box = True
-    set_box = True
-    unitcell = np.array([10., 11., 12.,
-                         0., 0., 0.,
-                         0., 0., 0.])
-    uni_args = (GRO,)
-
-    def test_unitcell_set2(self, ts):
-        box = np.array([80.017, 80.017, 80.017, 60.00, 60.00, 90.00],
-                       dtype=np.float32)
-
-        ref = np.array([80.00515747, 80.00515747, 56.57218552,  # v1x, v2y, v3z
-                        0., 0.,  # v1y v1z
-                        0., 0.,  # v2x v2y
-                        40.00257874, 40.00257874], dtype=np.float32)  # v3x, v3y
-        ts.dimensions = box
-        assert_almost_equal(ts._unitcell, ref, decimal=2)
 
 def test_multiframe_gro():
     u = mda.Universe(two_water_gro_multiframe)
