+++ testsuite/MDAnalysisTests/coordinates/test_netcdf.py
@@ -200,9 +200,8 @@ class TestNCDFReader2(object):
         assert_almost_equal(ref, u.trajectory.ts.dt, self.prec)
 
     def test_box(self, u):
-        ref = np.array([0., 0., 0., 0., 0., 0.], dtype=np.float32)
         for ts in u.trajectory:
-            assert_equal(ref, ts.dimensions)
+            assert ts.dimensions is None
 
 
 class TestNCDFReader3(object):
